requirements.txt
# Zenvi Backend - API Server
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
websockets>=13.0

# AI / LLM
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-ollama>=0.2.0

# AI Providers (media analysis)
openai>=1.50.0
google-generativeai>=0.8.0
google-cloud-vision>=3.7.0
boto3>=1.35.0  # AWS Rekognition

# Video indexing
twelvelabs>=0.3.0

# Video generation
runware>=0.4.0
httpx>=0.27.0
requests>=2.31.0

# Utilities
aiofiles>=23.0
